Fix mistmatch in projection return types
The type annotation on GlobalT.project is wrong, because projection with
merging is a partial function and may return None. This case was not
handled in many places where projection failure is not currently tested
by examples.

Fix by correcting the type annotations and adding the necessary checks
and early returns in projection implementations.